requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.3
CacheControl==0.13.1
cachetools==5.3.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==41.0.7
Django==4.2.11
django-crontab==0.7.1
django-environ==0.11.2
djangorestframework==3.14.0
firebase-admin==6.4.0
google-api-core==2.11.1
google-api-python-client==2.102.0
google-auth==2.27.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.26.0
google-cloud-storage==2.10.0
google-crc32c==1.8.0
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
grpcio==1.60.1
grpcio-status==1.60.1
gunicorn==21.2.0
httplib2==0.22.0
idna==3.6
joblib==1.3.2
msgpack==1.0.7
numpy==1.26.4
proto-plus==1.23.0
protobuf==4.25.3
psycopg2==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.12.1
pyparsing==3.1.1
python-dotenv==1.0.0
pytz==2024.1
requests==2.31.0
rsa==4.9
scikit-learn==1.3.2
scipy==1.11.4
soupsieve==2.4.1
sqlparse==0.4.4
threadpoolctl==3.2.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.1.0
whitenoise==6.6.0
